seperated components from route
