--- src/components/sortBy.tsx
@@ -1,48 +1,0 @@
-import { List } from "@/pages/list";
-
-export function sortByAlphabet(input: List) {
-  const sorted = [...input];
-  sorted.sort((a, b) => {
-    const nameA = a.item.toUpperCase();
-    const nameB = b.item.toUpperCase();
-    if (nameA < nameB) {
-      return -1;
-    }
-    if (nameA > nameB) {
-      return 1;
-    }
-    return 0;
-  });
-  return sorted;
-}
-export function sortByCategory(input: List): List {
-  const sorted = [...input];
-  sorted.sort((a, b) => {
-    const categoryA = a.category.toUpperCase();
-    const categoryB = b.category.toUpperCase();
-
-    const nameA = a.item.toUpperCase();
-    const nameB = b.item.toUpperCase();
-
-    if (categoryA < categoryB) {
-      return -1;
-    }
-    if (categoryA > categoryB) {
-      return 1;
-    }
-
-    if (categoryA === categoryB) {
-      if (nameA < nameB) {
-        return -1;
-      }
-      if (nameA > nameB) {
-        return 1;
-      }
-      return 0;
-    }
-    return 0;
-  });
-  return sorted;
-}
-
-export function sortByDate() {}
